Update section 10 Synopsis
I realized that I was writing in section 6 the specification that
belongs in section 10. I've rearranged things so that section 10 now has
an updated API outline (which should maybe be removed from section 6)
and I've spec'd `execution::nest()`, `execution::async_scope<>`, and
`execution::spawn()`.

I need to revisit section 6 to be a prose description of the tools I'm
spec'ing, and finish section 10.