More section 5 updates
Tweak the definitions in 5.1 to include a spawn-env and future-env to
make it easier to talk about the environment provided to the nested
senders by the receivers to which they're connected.

Reorganize section 5.4 to separate the environment-related discussion
from the lifecycle of the spawned operation's operation-state.

Write most of section 5.5, with a couple of TODOs left behind. Spent
some time contemplating whether the double-nest that Unifex performs is
necessary and decided against it. The spawned sender *is* nested to
ensure that any work it does can be protected from use-after-free by the
scope, but the future-sender is not nested; I should probably document
these thoughts in the paper to highlight the existence of the choice and
to explain why I chose what I did.