Add counting_scope::close() and simple_counting_scope (#12)
* Add counting_scope::close() and simple_counting_scope

Incorporate feedback from the folks working on P2300 and make the
following changes:
 - give `counting_scope` a `.close()` method, which makes `.join()` not
   responsible for closing the scope
 - give `counting_scope` an internal stop source and a public
   `request_stop()` method, like `unifex::v1::async_scope`
 - introduce `simple_counting_scope`, which is like `counting_scope` but
   without the stop source or the `.request_stop()` method

This first change only adds a call to `close()` in the rsys-based
example and adds the new type and methods to section 5.1 Definitions;
more change will come in this branch.

* Describe simple_counting_scope

This diff updates section 5.8 to describe `simple_counting_scope`
instead of `counting_scope`. We'll refer to section 5.8 when describing
`counting_scope` in section 5.9.

* Describe counting_scope in terms of simple_counting_scope

This diff gives a brief description of `counting_scope` in terms of
`simple_counting_scope` and `inplace_stop_source`. I'm breaking for
dinner here; I plan to add method-by-method descriptions.

* Finish section 5

This diff describes the methods of `counting_scope`, adds a little more
context to each state in the state machine diagram, and adds the missing
description for `simple_counting_scope::close()`.

* PR feedback