Update main to P3149r3 as published (#16)
* Remove counting_scope::joined(), counting_scope::join_started(), counting_scope::use_count()

* Update R2 description

Tweak the straw-poll markup to render better as a table and add
backticks around the names of the removed methods.

* Update paper number to D3149R2

* Update the description of counting_scope::nest()

Update `counting_scope::nest()` to explain when the scope's count of
outstanding senders is decremented.

* Add poll notes

* Fix poll table formatting

* Add a fourth state to counting_scope

* Update formatting

* Finalize P3149R2

 - Tweak the formatting of the revision history for R2
 - Ensure that `counting_scope::nest()` still provides the Strong
   Exception Guarantee in the face of the new "unused" state
 - Make some minor editorial tweaks to clarify wording now that the
   state machine is more complicated

* Update paper to R3 and make async_scope examples exception safe

* Remove merge conflict cruft

* Add example based on rsys

* Address comments

* Update asyncscope.md

Co-authored-by: Ian Petersen <ispeters@meta.com>

* Update asyncscope.md

Co-authored-by: Ian Petersen <ispeters@meta.com>

* Fix listener loop in HTTP server example

* Fix indentation

* Split async_scope into async_scope and async_scope_token; add let_with_async_scope (#9)

* Split the concepts, add let_with_async_scope decl

* Small updates to refined list of concepts

* Finish (I think) splitting the concepts

I think this diff completes the process of splitting the concepts; I
need to double-check that references to things like `nest(snd, scope)`
are all updated to `nest(snd, token)`, though.

* More edits

I'm making a pass over the document; I've updated whitespace, tried to
ensure APIs are used correctly, used tokens where appropriate, etc.

* Finish splitting scopes and tokens

* Address @jesswong's feedback

* Fix exception code to use let_error with counting_scope spawn

* Counting scope vs let with async (#11)

* Add blurb about counting_scope vs let_with_async_scope

* fix type

* Fix syntax bug

* Fix typos

* Reference P3296R0 for let_with_async_scope

* Fix formatting

* Resolve some PR feedback

Simplify part of the Tony Table comparing "old" code, code with a
`counting_scope`, and code with `let_with_async_scope`.

* Add manual citation for P3296

---------

Co-authored-by: Ian Petersen <ispeters@meta.com>

* Add examples for recursively spawned work using let_with_async_scope and counting_scope (#10)

* Add example for recursively spawned work using let_with_async_scope and counting_scope

* Update r3 info

* Fix wording

* Address comments

* Address comments and add spawn_future example with error surfacing

* Fix formatting

* Fix formatting

* PR feedback and formatting

---------

Co-authored-by: Ian Petersen <ispeters@meta.com>

* Add counting_scope::close() and simple_counting_scope (#12)

* Add counting_scope::close() and simple_counting_scope

Incorporate feedback from the folks working on P2300 and make the
following changes:
 - give `counting_scope` a `.close()` method, which makes `.join()` not
   responsible for closing the scope
 - give `counting_scope` an internal stop source and a public
   `request_stop()` method, like `unifex::v1::async_scope`
 - introduce `simple_counting_scope`, which is like `counting_scope` but
   without the stop source or the `.request_stop()` method

This first change only adds a call to `close()` in the rsys-based
example and adds the new type and methods to section 5.1 Definitions;
more change will come in this branch.

* Describe simple_counting_scope

This diff updates section 5.8 to describe `simple_counting_scope`
instead of `counting_scope`. We'll refer to section 5.8 when describing
`counting_scope` in section 5.9.

* Describe counting_scope in terms of simple_counting_scope

This diff gives a brief description of `counting_scope` in terms of
`simple_counting_scope` and `inplace_stop_source`. I'm breaking for
dinner here; I plan to add method-by-method descriptions.

* Finish section 5

This diff describes the methods of `counting_scope`, adds a little more
context to each state in the state machine diagram, and adds the missing
description for `simple_counting_scope::close()`.

* PR feedback

* Last-minute tweaks

* Update document number to P3149R3

* Fix formatting nits

* Replace counting_scope recursive spawn example without unifex::let_value_with

---------

Co-authored-by: Jessica Wong (Seattle) <jesswong@fb.com>