Change paper number and start changelog (#32)
* Change paper number and start changelog

Update paper number to D3149R6 and add the first R6 changelog entry.

* First pass at moving to try_associate()

We updated much of section 5 to move from scope tokens having a `nest()`
method to having a `try_associate() -> bool` method and a corresponding
`dissociate() -> void()` method.

We discovered that there's no way to implement `counting_scope`'s
"cancel all outstanding tasks" behaviour because there's no way to
associate the sender with the scope's stop token.

We're contemplating `std::optional<sender auto> try_associate(sender
auto&&)`. It'd be nice if `simple_counting_scope` could return an
optional containing  reference to the provided sender to avoid
unnecessary copies/moves; `counting_scope` needs to provide an optional
containing a new sender value. Not sure if `std::optional<T&>` has made
it into C++26, and, if it has, whether `std::optional<T&&>` is a thing.

There are several (~5?) TODOs added to the paper, too.

* Update nest() description

We remembered we need a `token.wrap()` method; this diff updates the
description of `nest()` to use the new token API.

* Update the description of spawn()

Refresh the wording of `spawn()` to use the new token API.

* Update spawn_future

This change mostly updates spawn_future() to be described in terms of
try_associate(), wrap(), and dissociate(), with some TODOs left behind.

* Tweak the concrete scope token types

Add `wrap()` methods to the two concrete scope token types we're
proposing.

* Flesh out the description of async_scope_token

Address the TODOs in section 5.2 by explaining what an async_scope_token
is.

* Finish updating the description of nest()

* Tweak the description of spawn()

* Address the TODO in section 5.5 describing spawn_future

* Update the example nest-sender

We realized that, to provide the strong exception guarantee, we *must*
invoke `token.wrap()` before `token.try_associate()`. Furthermore,
although a `token.try_associate(sender) -> optional<sender>` looks
prettier, it is insufficient for implementing `spawn()` with the strong
exception guarantee because `connect` can throw, and `connect` would be
invoked _after_ `try_associate()` has modified the scope's state.

* Update section 5 with Strong Exception Guarantee

Update the descriptions of the algorithms in section 5 to ensure they
provide the Strong Exception Guarantee. This mostly means invoking
`wrap()` before invoking `try_associate()`.

* Update parts of simple counting scope to use try_associate

* Update authors/contributors

Update Jessica and Ian's email addresses, and create a Contributors
section to reflect the fact that we've basically rewritten the whole
paper at this point.

* Small edits to simple_counting_scope

Make a few edits to @jesswong's updates to `simple_counting_scope`;
mostly just deleting left-over `TODO`s.

* Move the description of copying nest-senders

The unusual copying process for nest-senders used to be specific to the
scope that created the sender but it's now generic to all scopes so the
description needed to be moved to `execution::nest`.

* Minor tweaks to simple_counting_scope

* Progress on simple_counting_scope

Move the note about when a nest-sender's operation-state dissociates
from the scope into the description of nest(), and add an explanation of
simple_counting_scope::token::try_associate().

* Update to using async_scope_association

This iteration introduces the `async_scope_association<Assoc>` concept,
which is modeled by types that are RAII handles on associations with a
scope.

* let_async_scope updates

Change all instances of `let_with_async_scope` to `let_async_scope` to
match the actual proposed name. Update the paper reference to point to
the R2 version. Fix one or two cases where example uses of
`let_async_scope` had no predecessor sender.

* Formatting and counting_scope

Run most of the code through `clang-format` and finish the exposition on
how `counting_scope` works.

* Final updates for R6

- More fully describe the changes since R5
- Update sections 7 and 8 to match the new direction
- Add thanks to folks in the std::execution Discord

* Actual final commits

Update the paper number and correct a couple of oversights.

* Final final

Missed some cases where `try_associate()` was still returning `bool`. I
left the "true" and "false" words in the flow chart because they're
shorter than the alternatives and not *wrong*.

---------

Co-authored-by: jesswong <5247012+jesswong@users.noreply.github.com>