Update paper to R9 with suggested changes from LEWG (#47)
* Update paper to R9 with suggested changes from LEWG

* Address some missing similar changes

* remove empty space

* more nits

* nits

* Change decayed_tuple to decayed-tuple

* Clarify wrap() can't modify completions

* Add missing punctuation

Add a missing `*` when completing a receiver with `set_stopped` and
terminate a bullet point with a period.

* Rename _rcvr to r

* Fixup spawn-receiver; add receiver_concept, remove get_env

* Avoid storing an "associated" bool in spawn-state

Factor the "destroy and deallocate" logic out of spawn-state::complete
into a new expo-only method named destroy, and then call it from the
else branch in run when token.try_associate() fails.

* Fix up the spawn-future-receiver

* add the receiver_concept type alias
* ensure all the "tuples" are _decayed-tuple_
* make the try-catch blocks conditional on whether they're necessary

* Fixups to spawn_future

One more tuple -> decayed-tuple change, and address the fact that the
invocation of `visit` on the result variant was wrong: the argument to
the callable will be a decayed-tuple that needs to be decomposed into
its parts with std::apply (unless it's monostate...).

* Make sure we JOIN-ENV in the right order

JOIN-ENV(a, b) produces an environment that queries a before b so we
need to make sure we're constructing our joined environments in the
right order.

* JOIN-ENV appears to be a binary function

We were doing JOIN-ENV(a, b, c), but JOIN-ENV is binary so update to
JOIN-ENV(JOIN-ENV(a, b), c).

* Update section 5.5's description of spawn_future's stop_token

A previous commit updated the wording for how `spawn_future` computes a
stop token; this commit updates the corresponding prose in section 5.5.

* Delete the spawned opstate in the future op's dtor

Ensure that it's the future-sender's opstate that destroys the spawned
operation (via a call to abandon) when the future consumes the result of
the spawned work.

* Incorporate LWG feedback

A whole whack of changes from Wednesday and Friday sessions in LWG in
Hagenberg, Austria.

* Final tweaks

* Address @jesswong's PR feedback
* Reformat a few over-long lines
* Apply some LWG advice like "no braces around single-line if bodies"

* Update to a P-number

---------

Co-authored-by: Ian Petersen <ispeters@meta.com>