Update paper to P3149R10 with small-group wording feedback (#48)
* Update paper number to D3149R10

* Address Jen's comments

* Replace nest-data{} initialization with parenthesis
Replace nest-data{} with parentheses

* Reflow some long lines to be <=120 chars

This diff just reformats some long lines to fit in 120 columns or less.

* Mostly replace nest-data spec with prose

* Incorporate feedback from 3/28 session

This change incorporates changes to the specification of _`nest-data`_
suggested during a small group session with @dietmarkuehl, @jesswong,
and @mhoemmen.

* Address the rest of Jen's comments

* Update get-state to be noexcept(false)

* Partial application of small-group feedback

I'm pushing this midway through applying feedback from the last
small-group session to share the TODOs captured in the section of the
changelog relating to R10.

* Flesh out nest's get-state

This diff is the result of paring with @jesswong on implementing
`impls-for<nest_t>::get-state` in terms of the newly-defined `release()`
member function on _`nest-data`_. We *think* `get-state` is now correct
and exception-safe; we're still missing an articulation of when
`get-state` is `noexcept`, which is marked in the doc as a TODO.

* Add noexcept clause for nest_t::get_state

* Address most feedback from 5/9

This change addresses most of the feedback from the small-group session
on May 9, 2025 and leaves TODOs for what has yet to be addressed.

* Last-minute changes before mailing deadline

* Incorporate some of Eric Niebler's feedback from the Discord
* Update the revision log
* Attempt to reword [exec.spawn.future] paragraph 4 to better describe
  _`variant_t`_
* Other minor editorial changes

* Update paper number to P3149R10

R10 has been published so the in-doc number should start with P.

---------

Co-authored-by: jesswong <5247012+jesswong@users.noreply.github.com>