A better nest-data implementation
This diff introduces what I think is a good interface for the
association-enhanced _`nest-data`_. The central change is that
_`nest-data`_ now has this method:

```cpp
using scope_ref = unique_ptr<wrap_sender, deleter>;

pair<assoc_t, scope_ref> release() && noexcept;
```

where _`deleter`_ is a deleter that invokes the destructor on the
to-be-deleted pointer when it's not-null. This gives us a way to move
the association out of the source _`nest-data`_ and a reference to its
sender that we can also move from, with an attached guarantee that the
source sender will have its destructor invoked in all cases.

The one weird thing about this interface is that the returned `pair` is
only valid so long as the source _`nest-data`_ instance still exists
since the `scope_ref`'s destructor will reach into the source instance
to destroy the sender therein, which is clearly not allowed if the
source instance has already been destroyed.