Wednesday LWG changes
This diff incorporates the feedback we received during the final session
of LWG on Wednesday in Sofia. I've also tried to proactively update the
wording for `counting_scope` to match the feedback given on
`simple_counting_scope`, and made some editorial changes like updating
references to P2300 to refer to R9 and R10 where appropriate rather than
R7, added to the Acknowledgments, and a few other similar changes.